refactor: ViewController에서 View와 관련된 코드 분리
